RCA update InstitutionServiceTest and fix delete method of InstitutionService

+++ src/main/java/com/iesfranciscodelosrios/service/InstitutionService.java
@@ -27,6 +27,7 @@ public class InstitutionService implements iServices<Institution> {
 
     @Override
     public Institution delete(Institution institution) {
+        if (institution == null) return null;
         institutionRepository.delete(institution);
         return institution;
     }
